+++ prism4j-bundler/src/main/java/ru/noties/prism4j/bundler/LanguageInfo.java
@@ -41,4 +41,11 @@ public class LanguageInfo {
     public int hashCode() {
         return name.hashCode();
     }
+
+    @Override
+    public String toString() {
+        return "LanguageInfo{" +
+                "name='" + name + '\'' +
+                '}';
+    }
 }