+++ prism4j-bundler/CHANGELOG.md
@@ -1,3 +1,11 @@
-# 1.0.1
+# 1.1.0
+* Add `languages()` method to generated `GrammarLocator` class
+* Allow `includeAll` option (include all available languages)
+* Add language support:
+* * `groovy` (no string interpolation)
+* * `markdown`
+* * `scala`
+* * `swift`
 
+# 1.0.1
 * Fix for `javascript` and `yaml` grammar definitions to work on Android (escape `}` char in regex's)